# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
    "/api/v1/tasks/{id}" : {
      "get" : {
        "tags" : [ "task-controller" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "task-controller" ],
        "operationId" : "update",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "task-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/tasks/{id}/series" : {
      "put" : {
        "tags" : [ "task-controller" ],
        "operationId" : "applySeriesEdit",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TaskSeriesEditRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/holidays/{id}" : {
      "get" : {
        "tags" : [ "holiday-controller" ],
        "operationId" : "findById_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HolidayView"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "holiday-controller" ],
        "operationId" : "update_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateHolidayRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HolidayView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "holiday-controller" ],
        "operationId" : "delete_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/events/{id}" : {
      "get" : {
        "tags" : [ "event-controller" ],
        "operationId" : "findById_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventView"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "event-controller" ],
        "operationId" : "update_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EventView"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "event-controller" ],
        "operationId" : "delete_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/tasks" : {
      "get" : {
        "tags" : [ "task-controller" ],
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "schoolId",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TaskView"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "task-controller" ],
        "operationId" : "create",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TaskView"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks/{id}/status" : {
      "patch" : {
        "tags" : [ "task-controller" ],
        "operationId" : "updateStatus",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTaskStatusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskView"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "CreateTaskRequest" : {
        "required" : [ "assigneeUserIds", "dueDate", "schoolId", "title" ],
        "type" : "object",
        "properties" : {
          "title" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "schoolId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "classroomId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "assigneeUserIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          "dueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "dueTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "TODO", "IN_PROGRESS", "DONE" ]
          },
          "priority" : {
            "type" : "string",
            "enum" : [ "LOW", "MEDIUM", "HIGH" ]
          },
          "recurrence" : {
            "$ref" : "#/components/schemas/RecurrenceSpec"
          }
        }
      },
      "LocalTime" : {
        "type" : "object",
        "properties" : {
          "hour" : {
            "type" : "integer",
            "format" : "int32"
          },
          "minute" : {
            "type" : "integer",
            "format" : "int32"
          },
          "second" : {
            "type" : "integer",
            "format" : "int32"
          },
          "nano" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "RecurrenceSpec" : {
        "required" : [ "cycle", "untilDate" ],
        "type" : "object",
        "properties" : {
          "cycle" : {
            "type" : "string",
            "enum" : [ "DAILY", "WEEKLY", "MONTHLY" ]
          },
          "dueDayOfWeek" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dueDayOfMonth" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dueTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "untilDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "TaskView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "schoolId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "orgId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "classroomId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "assigneeUserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "dueDate" : {
            "type" : "string",
            "format" : "date"
          },
          "dueTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "TODO", "IN_PROGRESS", "DONE" ]
          },
          "priority" : {
            "type" : "string",
            "enum" : [ "LOW", "MEDIUM", "HIGH" ]
          },
          "recurrenceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "parentTaskGroupId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "TaskSeriesEditRequest" : {
        "required" : [ "choice", "occurrenceDate" ],
        "type" : "object",
        "properties" : {
          "choice" : {
            "type" : "string",
            "enum" : [ "JUST_THIS", "THIS_AND_FOLLOWING", "ENTIRE_SERIES" ]
          },
          "occurrenceDate" : {
            "type" : "string",
            "format" : "date"
          },
          "title" : {
            "type" : "string"
          },
          "dueTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "TODO", "IN_PROGRESS", "DONE" ]
          },
          "skipped" : {
            "type" : "boolean"
          }
        }
      },
      "CreateHolidayRequest" : {
        "required" : [ "date", "name", "schoolId" ],
        "type" : "object",
        "properties" : {
          "schoolId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "name" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          }
        }
      },
      "HolidayView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "schoolId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "name" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "source" : {
            "type" : "string",
            "enum" : [ "CUSTOM", "FEDERAL" ]
          },
          "approved" : {
            "type" : "boolean"
          },
          "approvedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "approvedByUserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "CreateEventRequest" : {
        "required" : [ "endDt", "schoolId", "startDt", "title", "type" ],
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "enum" : [ "CLASSROOM", "CUSTOM", "SCHOOL" ]
          },
          "schoolId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "classroomId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "startDt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endDt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "allDay" : {
            "type" : "boolean"
          },
          "organizerUserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "inviteParents" : {
            "type" : "boolean"
          },
          "attendeeUserIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          "studentIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          "excludedParticipantIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }
      },
      "EventView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "CLASSROOM", "CUSTOM", "SCHOOL" ]
          },
          "schoolId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "orgId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "classroomId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "startDt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "endDt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "allDay" : {
            "type" : "boolean"
          },
          "organizerUserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "inviteParents" : {
            "type" : "boolean"
          },
          "createdByUserId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "attendeeUserIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          "studentIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          },
          "excludedParticipants" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExcludedParticipantView"
            }
          },
          "softFlags" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/SoftFlagView"
            }
          }
        }
      },
      "ExcludedParticipantView" : {
        "type" : "object",
        "properties" : {
          "participantId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "participantType" : {
            "type" : "string"
          }
        }
      },
      "SoftFlagView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "conflictType" : {
            "type" : "string"
          },
          "conflictingEntityId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "UpdateTaskStatusRequest" : {
        "required" : [ "status" ],
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "TODO", "IN_PROGRESS", "DONE" ]
          }
        }
      },
...
    }
  }
}